+++ seed/template/multi-tic-tac-toe/angularjs/compiler/babeljs/mocha/components/box/main.spec.js
@@ -1,37 +1,35 @@
 describe('Component: box', () => {
     beforeEach(module('speedseed'))
 
+    let scope
     let ssBox
     let ssBoxCtrl
     let ssTable
+
     beforeEach(inject(($rootScope, $compile) => {
-        const scope = $rootScope.$new()
+        scope = $rootScope.$new()
 
         ssTable = angular.element('<ss-table></ss-table>')
         ssTable = $compile(ssTable)(scope)
-
-        scope.$apply()
     }))
 
     beforeEach(inject(($rootScope, $compile) => {
-        const scope = $rootScope.$new()
         ssBox = angular.element(ssTable.find('ss-box')[0])
         ssBox = $compile(ssBox)(scope)
 
-        scope.$apply()
-
         ssBoxCtrl = ssBox.controller('ssBox')
+
+        scope.$apply()
     }))
 
     const pushed = (param) => {
-        expect(ssBoxCtrl.xo)
-        .toBe(param)
+        assert.equal(ssBoxCtrl.xo, param)
 
         ssBoxCtrl.push()
     }
 
-    it('value of xo when pushed', () => {
-        pushed()
+    it('value global of xo when pushed', () => {
+        pushed(undefined)
         pushed('O')
         pushed('X')
         pushed('O')